task: very rough mobile nav
