--- src/app/(landing)/layout.module.scss
@@ -1,18 +1,0 @@
-@use '../../styles/mixins';
-@use '../../styles/colors';
-
-.container {
-  position: sticky;
-  z-index: 2;
-  top: 0;
-}
-
-.nav {
-  padding: mixins.spacing(x-big) 0;
-}
-
-.banner {
-  margin: -1 * mixins.spacing(x-big) 0 auto;
-  overflow-y: visible;
-  height: 1px;
-}
