+++ src/components/Button/Button.module.scss
@@ -8,6 +8,7 @@
   border-width: 0.0625rem;
   border-radius: 1000rem;
   display: inline-flex;
+  justify-content: center;
   border-style: solid;
   cursor: pointer;
