+++ src/styles/_variables.scss
@@ -1,6 +1,6 @@
 $media: (
   mobile: 425px,
-  tablet: 768px,
+  tablet: 820px,
   laptop: 1024px,
 );
